requirements.txt
fastapi
uvicorn
pillow
numpy==1.26.4
tflite-runtime==2.14
python-multipart